+++ include/BT/control.hpp
@@ -1,27 +1,9 @@
 #pragma once
 
+#include "node.hpp"
 #include <gtest/gtest.h>
-#include <string>
 
 namespace BT {
-enum Status {
-  SUCCESS,
-  FAILURE,
-  RUNNING,
-};
-
-class Node {
-protected:
-  Status state;
-  const std::string name;
-
-public:
-  Node(const std::string &name);
-  virtual Status tick();
-  std::string getName() const;
-  Status status() const;
-};
-
 class ControlNode : public Node {
 protected:
   int counter;
@@ -39,7 +21,7 @@ public:
 class Sequence final : public ControlNode {
 public:
   Sequence(const std::string &name, Node **children, int length);
-  Status tick() override final;
+  Status tick() override;
 
   FRIEND_TEST(SequenceTest, Tick);
 };
@@ -47,7 +29,7 @@ public:
 class Selector final : public ControlNode {
 public:
   Selector(const std::string &name, Node **children, int length);
-  Status tick() override final;
+  Status tick() override;
 
   FRIEND_TEST(SelectorTest, Tick);
 };
@@ -58,20 +40,6 @@ private:
 
 public:
   Parallel(const std::string &name, Node **children, int length, int threshold);
-  Status tick() override final;
-};
-
-class Decorator: public Node {
-protected:
-  Node *child;
-
-public:
-  Decorator(const std::string &name, Node *child);
-};
-
-class Inverter final: public Decorator {
-public:
-  Inverter(const std::string &name, Node *child);
-  Status tick() override final;
+  Status tick() override;
 };
 } // namespace BT
