+++ test/helper.hpp
@@ -1,0 +1,10 @@
+#include <BT/node.hpp>
+
+class ControlTestTick : public BT::Node {
+public:
+  ControlTestTick(const std::string &name, BT::Status status) : Node(name) {
+    state = status;
+  }
+
+  BT::Status tick() override { return state; }
+};
